refactor(config): make config a foundational leaf crate
ConfigYaml carried `Options` (re-exported from hplugin) and `FuseMode`
(from hdriver-bridge), so the pre-engine config crate transitively pulled
hplugin's tree (async-trait, erased-serde, futures, tokio) — defeating the
point of a lightweight crate that runs before the engine boots.

Both types are config by nature (their own docs say so), so move them down:
- `Options` / `decode_opt` / `deny_unknown` now live in `config`; `hplugin::config`
  re-exports them so every plugin author's import keeps resolving.
- `FuseMode` now lives in `config`; `hdriver_bridge::fuseconfig` re-exports it.

`config` drops its hplugin + hdriver-bridge deps and is now a leaf
(anyhow/memoize/serde/serde_yaml/tracing). hplugin and hdriver-bridge depend on
it instead. No call sites change — the re-exports preserve every path.
